+++ src/index.ts
@@ -1,10 +1,8 @@
 export { buildThaiAddressIndex } from './core/indexer'
-export { clearDefaultIndex } from './data/loader'
 export { searchThaiAddress } from './core/search'
 export { formatThaiAddressSuggestion } from './core/formatter'
 export { resolveThaiAddress } from './core/resolver'
 export { normalizeThaiAddressText } from './core/normalizer'
-export { useThaiAddressAutocomplete } from './react/useThaiAddressAutocomplete'
 
 export type {
   ThaiAddressRecord,
@@ -12,7 +10,7 @@ export type {
   ThaiAddressSuggestion,
   ResolvedThaiAddress,
   SearchOptions,
-  UseThaiAddressAutocompleteOptions,
+  BuildIndexOptions,
   RawData,
   RawGeography,
   RawProvince,